Protected user routes

+++ src/api/user_router.py
@@ -36,24 +36,35 @@ def get_user(
 
 
 @router.delete("/user/{email}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_user(email: EmailStr, db: SessionLocal = Depends(get_db)):
-    existing_user = UserService.get_user_by_email(db=db, email=email)
-    if not existing_user:
-        raise HTTPException(
-            status_code=404, detail=f"User not found for email: {email}"
-        )
-    existing_user = UserService.delete_user(db=db, user=existing_user)
-
-
-@router.put("/user/{email}", response_model=UserResponse)
-def update_user(
-    email: EmailStr, updated_fields: UserUpdate, db: SessionLocal = Depends(get_db)
+def delete_user(
+    email: EmailStr,
+    token: Annotated[UserToken, Depends(AuthUtils.get_token_from_header)],
+    db: SessionLocal = Depends(get_db),
 ):
     existing_user = UserService.get_user_by_email(db=db, email=email)
     if not existing_user:
         raise HTTPException(
             status_code=404, detail=f"User not found for email: {email}"
         )
+    if not existing_user.validate_jwt(token):
+        raise HTTPException(status_code=403, detail=f"User not authorized to resource.")
+    existing_user = UserService.delete_user(db=db, user=existing_user)
+
+
+@router.put("/user/{email}", response_model=UserResponse)
+def update_user(
+    email: EmailStr,
+    updated_fields: UserUpdate,
+    token: Annotated[UserToken, Depends(AuthUtils.get_token_from_header)],
+    db: SessionLocal = Depends(get_db),
+):
+    existing_user = UserService.get_user_by_email(db=db, email=email)
+    if not existing_user:
+        raise HTTPException(
+            status_code=404, detail=f"User not found for email: {email}"
+        )
+    if not existing_user.validate_jwt(token):
+        raise HTTPException(status_code=403, detail=f"User not authorized to resource.")
     return UserService.update_user(
         db=db, updated_fields=updated_fields, existing_user=existing_user
     )
